Used dependency graph to check the query reference
