+++ frontend/src/AppBuilder/Widgets/NewTable/_components/Footer/_components/ControlButtons.jsx
@@ -5,12 +5,9 @@ import { OverlayTriggerComponent } from './OverlayTriggerComponent';
 import useTableStore from '../../../_stores/tableStore';
 import { shallow } from 'zustand/shallow';
 import Popover from 'react-bootstrap/Popover';
-import { toast } from 'react-hot-toast';
 import { exportToCSV, exportToExcel, exportToPDF } from '@/AppBuilder/Widgets/NewTable/_utils/exportData';
 import { generateCypressDataCy } from '@/modules/common/helpers/cypressHelpers';
-import useStore from '@/AppBuilder/_stores/store';
-import { useModuleContext } from '@/AppBuilder/_contexts/ModuleContext';
-import { extractAndReplaceReferencesFromString } from '@/AppBuilder/_stores/utils';
+import { useTableRefresh } from '@/AppBuilder/Widgets/NewTable/_hooks/useTableRefresh';
 
 export const ControlButtons = memo(
   ({ id, table, darkMode, height, componentName, showAddNewRowPopup, setShowAddNewRowPopup, fireEvent }) => {
@@ -23,20 +20,7 @@ export const ControlButtons = memo(
     );
     const clientSidePagination = useTableStore((state) => state.getTableProperties(id)?.clientSidePagination, shallow);
     const hasDownloadEvent = useTableStore((state) => state.getHasDownloadEvent(id), shallow);
-    const isRefreshing = useTableStore((state) => state.getIsRefreshing(id), shallow);
-    const setIsRefreshing = useTableStore((state) => state.setIsRefreshing, shallow);
-
-    const { moduleId } = useModuleContext();
-    const dataRawValue = useStore((state) => {
-      const props = state.getComponentDefinition(id, moduleId)?.component?.definition?.properties;
-      const selector = props?.dataSourceSelector?.value;
-      return selector && selector !== 'rawJson' ? selector : props?.data?.value;
-    }, shallow);
-    const componentNameIdMapping = useStore((state) => state.getComponentNameIdMapping(moduleId), shallow);
-    const queryNameIdMapping = useStore((state) => state.getQueryNameIdMapping(moduleId), shallow);
-    const runQuery = useStore((state) => state.queryPanel.runQuery);
-    const dataQueries = useStore((state) => state.dataQuery.queries.modules?.[moduleId] ?? [], shallow);
-    const currentMode = useStore((state) => state.modeStore?.modules?.[moduleId]?.currentMode ?? 'edit');
+    const { handleRefresh, isRefreshing } = useTableRefresh(id);
 
     const RenderButton = ({ icon, tooltipId, tooltipContent, className, label, fill, ...restProps }) => {
       return (
@@ -182,47 +166,6 @@ export const ControlButtons = memo(
           </span>
         </>
       );
-    };
-
-    const handleRefresh = () => {
-      if (isRefreshing) return;
-
-      const queriesToRun = [];
-      const seen = new Set();
-
-      if (typeof dataRawValue === 'string' && dataRawValue) {
-        const { allRefs = [] } = extractAndReplaceReferencesFromString(
-          dataRawValue,
-          componentNameIdMapping || {},
-          queryNameIdMapping || {}
-        );
-
-        console.log('Manish :', { allRefs, dataRawValue, type: typeof dataRawValue });
-
-        for (const ref of allRefs) {
-          if (ref?.entityType !== 'queries') continue;
-          const entityNameOrId = ref.entityNameOrId;
-          if (!entityNameOrId) continue;
-          const query = dataQueries.find((dq) => dq.id === entityNameOrId || dq.name === entityNameOrId);
-          if (!query || seen.has(query.id)) continue;
-          seen.add(query.id);
-          queriesToRun.push(query);
-        }
-      }
-
-      if (queriesToRun.length === 0) {
-        toast('No queries to refresh');
-        return;
-      }
-
-      setIsRefreshing(id, true);
-      const runPromises = queriesToRun.map((query) =>
-        runQuery(query.id, query.name, undefined, currentMode, {}, undefined, undefined, false, false, moduleId)
-      );
-
-      Promise.allSettled(runPromises).finally(() => {
-        setIsRefreshing(id, false);
-      });
     };
 
     const renderRefreshButton = () => {
